Finished project

+++ todolist/views.py
@@ -4,17 +4,20 @@ from todolist.models import TaskList
 from todolist.form import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 
+@login_required
 def todolist(request):
-
     if request.method=="POST":
         form = TaskForm(request.POST or None)
         if form.is_valid():
+            instance = form.save(commit=False)
+            instance.manager = request.user
             form.save()
         messages.success(request,("New Task Added!"))
         return redirect('todolist')
     else:
-        all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manager=request.user)
         paginator = Paginator(all_tasks, 5)
         page = request.GET.get('page')
         all_tasks = paginator.get_page(page)
@@ -23,17 +26,23 @@ def todolist(request):
 def index(request):
     return render(request, 'index.html', {})
 
+@login_required
 def contact(request):
     return render(request, 'contact.html', {})
 
 def about(request):
     return render(request, 'about.html', {})
 
+@login_required
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete()
+    if task.manager == request.user:
+        task.delete()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
     return redirect('todolist')
 
+@login_required
 def edit_task(request, task_id):
     if request.method=="POST":
         task = TaskList.objects.get(pk=task_id)
@@ -47,14 +56,22 @@ def edit_task(request, task_id):
 
         return render(request, 'edit.html', {'task_obj': task_obj})
 
+@login_required
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True
-    task.save()
+    if task.manager == request.user:
+        task.done = True
+        task.save()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
     return redirect('todolist')
 
+@login_required
 def pending_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = False
-    task.save()
+    if task.manager == request.user:
+        task.done = False
+        task.save()
+    else:
+        messages.error(request,("Access Restricted, You Are Not Allowed!"))
     return redirect('todolist')
